requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pandas==2.2.3
matplotlib==3.10.1
reportlab==4.3.1
python-dotenv==1.1.0
google-generativeai==0.8.4
pydantic==2.11.2
